current model on my machine
I've uploaded the script i used to train my current model for species prediction. I used a ResNet-18, and adding iterative data augmentation based on the recall splits of the most recent epoch. The trouble I am running into is related to the huge species uneven splits, the smaller species are getting overlooked. For now I am going to keep this result, and start working on phase 2 to predict features of the species after it selects a species.